Common image tests reused on STY & PGL integration tests
