+++ src/integration/pages/photoGalleryPage/crossPlatformTests.js
@@ -1,6 +1,11 @@
-import { runCommonCrossPlatformTests, runTimestampTests } from '../../common';
+import {
+  runCommonCrossPlatformTests,
+  runTimestampTests,
+  runImageTests,
+} from '../../common';
 
 export default service => {
   runCommonCrossPlatformTests(service);
   runTimestampTests();
+  runImageTests();
 };
