+++ src/integration/pages/storyPage/ampTests.js
@@ -6,18 +6,6 @@ export default service => {
   runAmpAnalyticsTests();
   runCrossPlatformTests(service);
 
-  describe('Lead image', () => {
-    it('I can see an image with a caption', () => {
-      const imageEl = document.querySelector('main figure amp-img');
-      expect(imageEl).toBeInTheDocument();
-
-      const imageCaptionEl = document.querySelector('main figure figcaption');
-      expect(imageCaptionEl).toBeInTheDocument();
-      expect(imageCaptionEl.textContent).toBeTruthy();
-      expect(imageCaptionEl.textContent).toMatchSnapshot();
-    });
-  });
-
   describe('Social Embeds', () => {
     const hasRichInstagramEmbed = !!document.querySelector(
       'iframe.instagram-media',
