+++ src/integration/pages/storyPage/crossPlatformTests.js
@@ -1,8 +1,13 @@
-import { runCommonCrossPlatformTests, runTimestampTests } from '../../common';
+import {
+  runCommonCrossPlatformTests,
+  runTimestampTests,
+  runImageTests,
+} from '../../common';
 
 export default service => {
   runCommonCrossPlatformTests(service);
   runTimestampTests();
+  runImageTests();
 
   describe('Social Embeds', () => {
     it('I can see the skip link', () => {
